requirements.txt
torch>=2.0.0
torchvision>=0.15.0
albumentations>=1.3.0
wandb>=0.15.0
tqdm>=4.65.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.3.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
pytorch-metric-learning>=2.3.0